+++ src/mongo/routes/AdminConsoleRoute.ts
@@ -15,6 +15,7 @@ const { Config, Inject } = ObjectDecorators;
 export class AdminConsoleRoute extends ReactRoute {
     protected readonly appDir: string = "apps/admin";
     protected readonly hydrate: boolean = true;
+    protected readonly router: boolean = true;
 
     @Inject(ObjectFactory)
     protected siteSettingsObjectFactory!: ObjectFactory;