+++ vite.config.ts
@@ -1,3 +1,3 @@
 import { createViteConfig } from "@rapidrest/react/vite";
 
-export default createViteConfig({ appDir: ["apps/www", "apps/admin"] });
+export default createViteConfig({ appDir: ["apps/www", "apps/admin"], router: true });